feat(builtins)!: PJXEmptyState collapses slots into {{ content }}
Remove presentational slots (image, title, description, action, actions);
add single freeform `content` field. Keep `suggestions` chips unchanged.
Trim CSS to container + chip rules only. Update golden snapshot.
