+++ pyjinhx/builtins/ui/pjx_empty_state/pjx-empty-state.html
@@ -1,16 +1,1 @@
-<div class="pjx-empty-state{% if class_name %} {{ class_name }}{% endif %}" id="{{ id }}">
-    {% if image %}
-    <div class="pjx-empty-state__image">{{ image }}</div>
-    {% endif %}
-    <h3 class="pjx-empty-state__title">{{ title }}</h3>
-    <p class="pjx-empty-state__desc">{{ description }}</p>
-    {% if action %}
-    <div class="pjx-empty-state__action">{{ action }}</div>
-    {% endif %}
-    {% if actions %}
-    <div class="pjx-empty-state__actions">{% for item in actions %}{{ item }}{% endfor %}</div>
-    {% endif %}
-    {% if suggestions %}
-    <div class="pjx-empty-state__suggestions">{% for chip in suggestions %}<button type="button" class="pjx-empty-state__chip" data-pjx-suggestion="{{ chip.value if chip.value is defined else chip.label }}" data-pjx-event="{{ chip.event if chip.event is defined else 'pjx:suggestion' }}" @click="$dispatch($el.dataset.pjxEvent, { value: $el.dataset.pjxSuggestion })">{{ chip.label }}</button>{% endfor %}</div>
-    {% endif %}
-</div>
+<div class="pjx-empty-state{% if class_name %} {{ class_name }}{% endif %}" id="{{ id }}">{{ content }}{% if suggestions %}<div class="pjx-empty-state__suggestions">{% for chip in suggestions %}<button type="button" class="pjx-empty-state__chip" data-pjx-suggestion="{{ chip.value if chip.value is defined else chip.label }}" data-pjx-event="{{ chip.event if chip.event is defined else 'pjx:suggestion' }}" @click="$dispatch($el.dataset.pjxEvent, { value: $el.dataset.pjxSuggestion })">{{ chip.label }}</button>{% endfor %}</div>{% endif %}</div>
