+++ pyjinhx/builtins/ui/pjx_empty_state/pjx_empty_state.py
@@ -1,24 +1,17 @@
-from typing import Annotated, Any
+from typing import Any
 
 from pydantic import Field
 
-from pyjinhx import BaseComponent, Slot
-from pyjinhx.base import AttrValue, ExtraAttrs, PjxSlot
+from pyjinhx import BaseComponent
+from pyjinhx.base import AttrValue, ExtraAttrs
 
 
 class PJXEmptyState(BaseComponent):
-    image: Slot = ""
-    title: str = ""  # heading text — escaped
-    description: str = ""  # body text — escaped
-    action: Slot = ""
-    actions: Annotated[list[str | BaseComponent], PjxSlot()] = Field(
-        default_factory=list
-    )
-    # First-class interactive suggestion chips (option a from issue #77).
-    # Each item is a dict with:
-    #   - label      (str): button text
-    #   - value      (str, optional): dispatched with the event; defaults to label
-    #   - event      (str, optional): custom event name; defaults to "pjx:suggestion"
+    # First-class interactive suggestion chips (issue #77). Each item is a dict:
+    #   - label (str): button text
+    #   - value (str, optional): dispatched with the event; defaults to label
+    #   - event (str, optional): custom event name; defaults to "pjx:suggestion"
     suggestions: list[Any] = Field(default_factory=list)
     class_name: AttrValue = ""
+    content: str | BaseComponent = ""
     extra_attrs: ExtraAttrs = Field(default_factory=dict)
