+++ tests/unit/golden/empty_state.html
@@ -1,10 +1,1 @@
-<div class="pjx-empty-state" id="g">
-    
-    <h3 class="pjx-empty-state__title">Nothing</h3>
-    <p class="pjx-empty-state__desc">D</p>
-    
-    <div class="pjx-empty-state__action"><button>A</button></div>
-    
-    
-    
-</div>
+<div class="pjx-empty-state" id="g"><h3>Nothing</h3><p>D</p><button>A</button></div>
